bin scripts moved to cli_manager
