+++ setup.cfg
@@ -78,7 +78,20 @@ testing =
 # And any other entry points, for example:
 # pyscaffold.cli =
 #     awesome = pyscaffoldext.awesome.extension:AwesomeExtension
-
+console_scripts =
+    aggregate_column_data = py_cmdtabs.cli_manager:aggregate_column_data
+    column_filter = py_cmdtabs.cli_manager:column_filter
+    create_metric_table = py_cmdtabs.cli_manager:create_metric_table
+    desaggregate_column_data = py_cmdtabs.cli_manager:desaggregate_column_data
+    excel_to_tabular = py_cmdtabs.cli_manager:excel_to_tabular
+    filter_by_list = py_cmdtabs.cli_manager:filter_by_list
+    intersect_columns = py_cmdtabs.cli_manager:intersect_columns
+    merge_tabular = py_cmdtabs.cli_manager:merge_tabular
+    records_count = py_cmdtabs.cli_manager:records_count
+    standard_name_replacer = py_cmdtabs.cli_manager:standard_name_replacer
+    table_linker = py_cmdtabs.cli_manager:table_linker
+    tag_table = py_cmdtabs.cli_manager:tag_table
+    
 [tool:pytest]
 # Specify command line options as you would do when invoking pytest directly.
 # e.g. --cov-report html (or xml) for html/xml output or --junitxml junit.xml
